Tweak graph title/legend labels. mark WIP/TODO to add GEOPHIRES profile data to graph
