+++ src/geophires_docs/__init__.py
@@ -1,10 +1,15 @@
 from __future__ import annotations
 
+import json
 import os
 from pathlib import Path
 from typing import Any
 
+from pint.facets.plain import PlainQuantity
+
 from geophires_x_client import GeophiresInputParameters
+from geophires_x_client import GeophiresXClient
+from geophires_x_client import GeophiresXResult
 
 
 def _get_file_path(file_name) -> Path:
@@ -75,3 +80,20 @@ def _get_input_parameters_dict(  # TODO consolidate with FervoProjectCape5TestCa
         # TODO preserve newlines
 
     return ret
+
+
+def _get_full_profile(input_and_result: tuple[GeophiresInputParameters, GeophiresXResult], profile_key: str):
+    input_params: GeophiresInputParameters = input_and_result[0]
+    result = GeophiresXClient().get_geophires_result(input_params)
+
+    with open(result.json_output_file_path, encoding='utf-8') as f:
+        full_result_obj = json.load(f)
+
+    net_gen_obj = full_result_obj[profile_key]
+    net_gen_obj_unit = net_gen_obj['CurrentUnits'].replace('CELSIUS', 'degC')
+    profile = [PlainQuantity(it, net_gen_obj_unit) for it in net_gen_obj['value']]
+    return profile
+
+
+def _get_full_production_temperature_profile(input_and_result: tuple[GeophiresInputParameters, GeophiresXResult]):
+    return _get_full_profile(input_and_result, 'Produced Temperature')
